requirements.txt
hydra-core
PyYAML
python-dotenv
pyrootutils
opencv-python-headless
numpy
openvino-dev[onnx]
matplotlib
colorlog
Pillow
scipy
ipyfilechooser
onnx
onnxruntime
requests
pandas
flask
psutil
python-socketio
eventlet
PyJWT
seaborn
plotly
dash
dash-bootstrap-components
